requirements.txt
aiohttp==3.7.4.post0
aioredis==1.3.1
appdirs==1.4.4
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.2.1
Automat==20.2.0
autopep8==1.5.7
bcrypt==3.2.0
branca==0.4.2
brotlipy==0.7.0
cachetools==4.2.2
certifi==2021.10.8
cffi==1.14.6
channels==3.0.4
channels-redis==3.3.0
chardet==4.0.0
charset-normalizer==2.0.0
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
dj-database-url==0.5.0
Django==3.2.4
django-cors-headers==3.7.0
django-environ==0.4.5
django-heroku==0.3.1
djangorestframework==3.12.4
earthengine-api==0.1.268
folium==0.0.0
future==0.18.2
google-api-core==1.31.2
google-api-python-client==1.12.8
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-cloud-core==1.7.1
google-cloud-storage==1.19.0
google-crc32c==1.1.2
google-resumable-media==2.0.2
googleapis-common-protos==1.53.0
grpcio==1.38.1
gunicorn==20.1.0
hiredis==2.0.0
httplib2==0.19.1
hyperlink==21.0.0
idna==3.1
incremental==21.3.0
Jinja2==3.0.1
MarkupSafe==2.0.1
msgpack==1.0.2
multidict==5.1.0
numpy==1.21.2
packaging==21.0
pip==21.2.4
protobuf==3.18.0
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.7.0
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyshp==2.1.3
PySocks==1.7.1
pytz==2021.1
pyu2f==0.1.5
requests==2.26.0
rsa==4.7.2
service-identity==18.1.0
setuptools==58.0.4
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
wheel==0.37.0
whitenoise==5.3.0
yarl==1.6.3
zope.interface==5.4.0
